fix: Change data class id to Long. Pass MyApp.INSTANCE inside JsonLoad. Remove DB insert return values. Remove JsonActor and JsonGenre classes. Remove tools:targetApi="lollipop" in themes.

+++ app/src/main/java/com/aacademy/homework/data/FakeDataRepository.kt
@@ -14,7 +14,7 @@ object FakeDataRepository {
             }
     }
 
-    suspend fun getMovieDetail(id: Int): MovieDetailWithActors = FakeLocalRepository.getMovieDetail(id)
+    suspend fun getMovieDetail(id: Long): MovieDetailWithActors = FakeLocalRepository.getMovieDetail(id)
         .let {
             if (it.isEmpty()) FakeApiRepository.getMovieDetail(id) else it.first()
         }
