+++ app/src/main/java/com/aacademy/homework/data/model/MovieActor.kt
@@ -4,6 +4,6 @@ import androidx.room.Entity
 
 @Entity(primaryKeys = ["movieId", "actorId"])
 data class MovieActor(
-    val movieId: Int,
-    val actorId: Int
+    val movieId: Long,
+    val actorId: Long
 )
